feat(shared): move errorFilter + request to frontend/src/shared; extract types to types/ and add types/api.ts
