+++ .gitignore
@@ -142,4 +142,5 @@ vite.config.ts.timestamp-*
 
 # Ignore local issue drafts folder
 .github/ISSUES/
+.github/PR/
 backend/dist
