+++ frontend/src/scripts/stats.ts
@@ -1,6 +1,6 @@
 import { el, setChildren } from 'redom';
-import type { ErrorItem, PeriodStats } from './utils/errors';
-import { request } from './utils/request';
+import type { ErrorItem, PeriodStats } from '../../../types/errors';
+import { request } from '../shared/request';
 import { qs, qsa, createElement } from './utils/dom';
 import { t, getLabel, onLangChange } from './utils/i18n';
 import { typeColors, statusColors } from './utils/colors';
